requirements.txt
flask
flask-cors
tensorflow==2.14.0
numpy
pillow
scikit-learn
gunicorn
gdown
